requirements.txt
asgiref==3.11.1
black==26.3.1
click==8.3.1
colorama==0.4.6
Django==6.0.3
djangorestframework==3.17.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pytokens==0.4.1
sqlparse==0.5.5
tzdata==2025.3
